backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
psycopg2-binary
pdfplumber
pytesseract
Pillow
pandas
openpyxl
groq
python-dotenv
pydantic
